ca298/requirements.txt
asgiref==3.3.1
certifi==2020.6.20
Django==3.1.5
django-crispy-forms==1.11.1
django-extensions==3.1.1
djangorestframework==3.12.2
graphviz==0.16
pydot==1.4.2
pyparsing==2.4.7
pytz==2020.5
sqlparse==0.4.1
wincertstore==0.2
django-cors-headers